requirements.txt
torch~=1.7.0+cu101
torchvision~=0.8.0
matplotlib~=3.4.3
numpy~=1.20.3
visdom~=0.1.8.9
opencv-python~=4.5.3.56
pillow~=8.3.2
pandas~=1.3.3
sklearn~=0.0
scikit-learn~=1.0.2